Use a JSON config file for defining the entites & attributes to codegen
The code generation tool uses an XML config file for defining the
entities and attributes to run code generation for. There are two lines;
one for entities, the other for the attributes for each entity. Given
the number of attributes, managing these on a single line is becoming
problematic. It also doesn't work well with diff & merging tools.

This modifies our script to use crm_attributes.json to populate the
two lines in the XML config file before we run code generation. As it's
a JSON file, it's much more tooling-friendly.