Introduce error codes
Adds a registry of errors, keyed by a unique integer error code. Each
error also has a Title and, optionally, Detail. The content for both of
these is sourced from the StringResources file.

Request handlers throw an `ErrorException` passing in an error code. A
controller action can add a `MapErrorAttribute` to 'catch' the error,
translate it to a ProblemDetails object and set the response code
appropriately.

A top-level exception filter catches all ErrorExceptions and returns a
ProblemDetails response with a 400 response code.

In future we will likely want to add instance-specific data to each
Error but I've omitted that for now.

Also adds a helper method for asserting one of these errors was thrown.