Automatically clean up integration test data
Previously tests were required to explicity track the entities created
for tests so that the cleanup process knew what to delete. This was an
easy thing to forget and for 'black box' testing a test may not even
know what was created. This change makes the tracking automatic.

A proxy around the underlying CRM SDK (IOrganizationServiceAsync2
specifically) intercepts calls to Create-like methods and holds onto the
IDs of successful calls. They can then be cleaned up when required. A
type - TestDataScope - manages all this; it creates a proxy object to
track created entities then deletes everything when DisposeAsync() is
called. In effect, the lifetime of all entities created via the scope is
tied to the lifetime of the scope itself.