Shorten CRM timeouts
We have several errors in our logs around timeouts, many of which are in
excess of 5 minutes. We don't want requests running for anywhere near
that long as many clients will timeout long before that.

This shortens the connection timeout and reduces the retry count and
pause between each retry in order that we fail faster.