Make authorization policy more configurable
Until now we've had a single authorization policy applied across the
whole application, using our API key-based authentication scheme. For V3
endpoints we need to be able to authorize using an access token from ID
for particular endpoints.

This removes the application-wide AuthorizeFilter and replaces it with a
convention; this convention applies the same policy we've always used
for v1 and v2 endpoints but expects v3 endpoints to explicitly declare
the policy they require.