Add AllowIdSignInWithProhibitions member to v3 teacher response
This is a short term solution so that ID can prevent sign in (or
registration) of prohibited teachers *unless* a flag has been set on
the matched DQT record.

As it's for our use only and short term, I've hidden it from the Swagger
doc and prefixed the include entry with an '_'.