Fix EventObserver.AssertEventsSaved to work with synced events
Normally events are created via EF and we use an interceptor in tests to
'capture' the events that were created during the running of a
particular test to later assert on. The TrsDataSyncHelper doesn't use EF
to sync events, however, so this interceptor doesn't ever see the
events.

This change amends TrsDataSyncHelper to provide an IObservable that
emits synced models. In our test base class we subscribe to that and
push events onto the CaptureEventObserver for the corresponding test.