Only use journey state for mutable data
Until now we've used journey state as a kind of cache for any data that
the journey needs. I've changed this around so that only mutable data is
stored.

The filter approach we started for 'personId' is now extended to cover
MQs and this moves to using strongly-typed features and extension
methods instead of HttpContext.Items.