Split Change log UI into partial views
We're about to add even more types of entries to the change log and it's
already somewhat complex. This introduces a partial view for each type
of change in the list. There's also support for handling events (with a
partial for each event type), though actually querying events isn't
currently wired up.