Add a script and just recipe to clear schema cache
When a test project deploys the DB, it stashes away a hash of the
current schema definition in an AppData folder. The next time the tests
are run, the hash is re-calculated and if it matches what's in the
AppData file we know the schema hasn't changed and we can skip
re-deploying the DB.

Occasionally something goes wrong and we want to force re-deploying the
DB. This script is a helpful shortcut for doing that.