Write all TRN request data to TRS DB (#1812)
We need to be able to return data submitted in the original TRN request
from our API endpoints but currently that data isn't easily available.
This is the first step to fixing that - stashing all data from all TRN
requests in the TRS DB. A separate PR will follow to back-fill.