Induction fix 'check your answers' (#1833)
### Context

Issues arising from the review of the implemented inductions wizard

### Changes proposed in this pull request

Add validation to the CYA page to stop the Invalidation of the start and
completed date by users using the browser back button-

https://trello.com/c/ARYIRGrR/764-induction-fix-additional-browser-back-button-fix

Improve the display of the selected exemption reasons in the CYA page
(and also in the Induction summary card as pointed out by AH)

https://trello.com/c/WpZQBftk/756-induction-fix-adjust-exemption-reasons-pattern-on-cya-page

The colour of the “Not provided” value should be black

https://trello.com/c/UVca4Pcx/757-induction-fix-adjust-contrast-on-not-provided-values-on-cya

A ticket that lists some current behaviour, but the new behaviour is
that
1. If the user goes from CYA back to change status, the pages that are
shown in the journey should be pre-populated with current values.
2. "User should always see all the teacher's induction details [on the
CYA page.] I.e. status should always be displayed no matter the entry
point to the journey."

https://trello.com/c/o0cOv8Zv/763-induction-fix-align-cya-with-design-system-guidance

On the edit status page, remove logic to hide the original status from
the radio buttons but don’t pre-select the status field. I.e. if the
current status is exempt, the user should see the exempt.
All fields, besides induction status should be prepopulated with the
previous information where applicable.
https://trello.com/c/AwMa6p94/759-induction-fix-prepopulate-fields