Remodel induction to handle removing exemptions (#1855)
When an induction exemption is removed (e.g. because QTLS has gone away)
we need to know what `InductionStatus` to set the person to. Previously,
this wasn't possible - we weren't storing enough information.

This change adds an `InductionStatusWithoutExemption` property to
`Person`. It contains the status that the person would have if they
weren't exempt. When the last induction exemption is removed, we revert
to this status. To make this work, we also need to leave the
`StartedDate` and `CompleteDate` values intact when we initially change
the status to `Exempt`.

This also removes the need for the separate `CpdInduction*` fields.