Handle exemptions without a reason (#1859)
We have lots of production data that has people with `InductionStatus`
of `Exempt` without any exemption reasons. Currently, if we were to add
another exemption reason then remove it we would incorrectly change the
status to `RequiredToComplete`.

This change adds a separate flag - `InductionExemptWithoutReason` - that
we check before undoing an `Exempt` status when removing a reason. If
the flag is set, we keep the status as `Exempt`. Whenever the status is
explicitly set, we clear this flag.