Handle not-yet-synced person in AddInductionExemptionMessageHandler (#1874)
We could get an `AddInductionExemptionMessage` before we've synced the
Person for new records; this ensures we've synced the person before we
try to set their induction status.

We're also getting messages without the `DqtUser*` fields in cases where
the API was the source of the message. This handles that by falling back
to `SystemUserId`.