Add Enddate Page for routes (#1893)
Add End date Page for routes

This can be reviewed and merged as it's under a feature flag. 
This PR includes unifying the two 'FieldRequirement' enums into one.
EndDate only shows on the details page if it's applicable to the route
and status combo, and the change link will always appear if EndDate is
applicable
EndDate page validation is kept simple - if you've asked to edit the end
date you have to add a value (removes the need for for
mandatory/optional checks)


https://educationgovuk.sharepoint.com.mcas.ms/:x:/s/TRATransformationTeamDocs/ETLJKljZyxJCig3Pz_sPH2sBMV8QMbXNBs0EZfpKxA9ZCg?e=SmjX7Z).

- Added page
- Added Tests


https://trello.com/c/bX0M9P80/837-routes-and-professional-status-front-end-edit-journey-end-date

---------

Co-authored-by: CLAWLOR <catherine.lawlor@education.gov.uk>