Added Person Email Address and Employer Postcode to TPS Employment table and import process (#1912)
### Context

Workforce data is being used by the Find Teachers For Research
application and a number of additional fields have been identified which
would improve this.

### Changes proposed in this pull request

Add a `member_email_address` and `employer_postcode` to the
`tps_employments` table in TRS and amend the import job to include these
fields.

Remember to include a migration for the reporting db.