Populate dfeta_allowpiiupdatesfromregister based on AllowContactPiiUpdatesFromUserIds config setting (#1924)
### Context

There are rules in place that control whether we accept PII updates from
Register over the API. Those rules rely on the value of the
dfeta_allowpiiupdatesfromregisterattribute on the contact. The new TRN
allocation API needs to be extended to populate that attribute.

### Changes proposed in this pull request

Populate the dfeta_allowpiiupdatesfromregister attribute when creating a
contact in the v3 TRN allocation endpoint. Its value should be true when
the calling user ID is present in the AllowContactPiiUpdatesFromUserIds
configuration array.