Added employer email to workforce data (#1927)
### Context

Workforce data is being used by the Find Teachers For Research
application and a number of additional fields have been identified which
would improve this.

### Changes proposed in this pull request

Add a `employer_email_address` to the `tps_employments` table in TRS and
amend the import job to include this field.

Add a backfill job to populate this field for the existing records in
the `tps_employments` table.

Remember to include a migration for the reporting db.