Amended workforce data import to trim whitespace from email addresses (#1930)
### Context

The workforce data extract has been amended to include (and validate)
the estblishment email address however when it was run in production it
appears that some of the data has whitespace for the email address
rather than just being empty - this is currently being rejected as
invalid but really it could be trimmed and treated as null.

### Changes proposed in this pull request

Amend the workforce data import to trim any whitespace from the
establishment_email_address before validating it (empty is allowed).
Applied the same logic for member_email_address too.