Add back-end for new user roles (#1932)
### Context

User roles in TRS are being changed from having multiple roles per user
to a single role, with a different set of permissions. This PR
implements the new roles in parallel to the existing ("legacy") roles so
that existing users will still have access. Once the DQT migration is
complete, existing users will be moved to the new role structure and the
legacy role structure deleted.

### Changes proposed in this pull request

* Creates `role` column in `users` table
* Renames existing `UserRoles` class to `LegacyUserRoles`
* Implements new `UserRoles` and `UserPermissions` classes - new roles
have granular permissions with each permission type having a
`UserPermissionLevel` value of `No`, `View`, `Edit`
* Refactors policies to use `IAuthorizationRequirement`s and
`AuthorizationHandler`s,
* Moves policy implementation using legacy roles to `Legacy*` handler
classes
* Creates new implementation of policies using new role/permission
structure with `AuthorizationHandler`s handling the same
`IAuthorizationRequirements` so that a user can be authorized using
either the legacy roles or the new role structure
* Changes tests to use the new role structure (except for tests around
the UI for editing users, these have to continue to use the legacy role
structure until the new UI is implemented)

### Checklist

-   [x] Attach to Trello card
-   [x] Rebased master
-   [x] Cleaned commit history
-   [x] Tested by running locally
-   [ ] Run DQT integration tests locally (if appropriate)