Routes edit status (#1931)
### Context

Edit the status of the route (e.g. in progress etc)

Editing status to awarded/approved results in up to 3 other pages
needing to be completed (depending on the route) - end date, award date,
induction exemption. If user doesn't get to the end of this data
collection, none of the data is stored.

Moving status from awarded/approved to something else means award date
and induction exemption needs to be reset.

### Changes proposed in this pull request


https://trello.com/c/jQqmrejk/905-routes-and-professional-status-front-end-edit-journey-status

### Other changes in this PR:
Cancel button an all the edit pages goes back to Qualification tab
Details page shows all fields, not just those associated with the
route/status combo
Check that all mandatory data is present before saving it - bounce back
to details page if not
Editing fields from the check-your-answers (CYA) page - continue goes to
CYA, except for changing status to awarded/approved which then shows all
the pages associated with completing the route (as described above).