Amended to default training country reference as XK in DQT when left null in API (#1954)
### Context

During testing of the new routes v3 API, a bug was found where ITT
records are created with null training country.

This is due to small piece of logic not being ported from the v2 API.

Any route types allowing null in the API need to default to a United
Kingdom country code (XK) in the DQT ITT record as it is mandatory and
will cause issues if records need to be amended manually.