Handle 'definite duplicate' TRN requests in GET endpoint (#1963)
During TRN allocation, when we've found a 'definite duplicate' (i.e.
exactly one record matches with NINO and DOB), we return its TRN
immediately with status `Completed`. However, if the `GET` endpoint were
subsequently polled, we'd return a `404` since we have no record of the
request in the `trn_requests` table (legacy) and we have no CRM
`contact` with matching `dfeta_trnrequestid`.

This change adds `ResolvedPersonId` to `TrnRequestMetadata` and
populates it during TRN allocation if the resolved record is known.
Every time a `Pending` request is queried we check if it's been
completed since the last time and populate `ResolvedPersonId` if it has
been.

At some point we'll need a back-fill job to populate this field for all
existing requests, since we won't have `dfeta_trnrequestid` available
after the contact record has been migrated to TRS.