Show differences between matched record and TRN request (#1991)
This adds highlights to matched records to show which attributes on the
matched `Person` are different from the TRN request.

I've repurposed the `OneLoginUserMatchedAttribute` enum and renamed it
to `PersonMatchedAttribute`.

The method that figures out the differences is currently on the
`Matches` `PageModel`; that'll likely move with the next PR so it can be
re-used on the 'merge' page. The logic here is different to what we use
for the matching; here we want to show all differences (even in case)
whereas the matching is more permissive. (e.g. we could have a match on
first name but still have differences between the two first names.)


https://trello.com/c/kvHCycNP/1055-add-ui-for-managing-api-trn-request-potential-duplicates

<img width="523" alt="image"
src="https://github.com/user-attachments/assets/f0014c03-485a-4a60-a5bb-5344981e9d8c"
/>