Choose attributes to merge from TRN request (#2002)
This adds the UI for selecting which attributes should be retained on
the selected record vs copied from the TRN request.

I've also fixed the Matches page to redirect the user to Check Answers
if 'I want to create a new record ...' is selected.

There a few rough edges here (mostly around content) but the snagging
session on the journey as a whole will address those.


![localhost_44321_support-tasks_api-trn-requests_DEV-1234_merge_ffiid=167f1298-6399-4daa-aef7-6045b64b0f4e](https://github.com/user-attachments/assets/bb948e78-192a-47b2-84f0-0c27fcd11e5d)