Add tag helper for generating a sortable-table (#2005)
MOJ design system includes a `sortable-table` component but it runs
client-side. We need it to work with pagination so we need it to run
server side. This adds a tag helper to aid with setting it up.

I've noticed we're referencing a very old version of moj-frontend; when
we update we'll likely need to update this tag helper.

A PR showing this in use is at
https://github.com/DFE-Digital/teaching-record-system/pull/2006