Add mechanism for creating a TransactionScope around an endpoint (#2012)
We have places where we want to enqueue a Hangfire job and update our DB
atomically. We can do this if both of those operations are within a
`TransactionScope`.

The challenge is we have a request-scoped `DbContext` so creating a
`TransactionScope` in a Razor page handler is too late (since the
`DbContext` may already be up and running for the request if it's been
used in a filter).

This adds a mechanism for decorating a Razor page with
`[TransactionScope]`; if that's present a `TransactionScope` will be
setup immediately after the route has been identified. With that, all
our filters and the page handlers will all share a `TransactionScope`.