Add TRN Request resolution details to SupportTask data (#2016)
This adds the resolved person attributes to the `SupportTask`s data. As
part of this I've had to fix up how we're updating `Data` (it's a bit
different since it's stored as JSON in the DB).

I've also pulled some of the common methods into a `PageModel` base
class for the journey.