Add an interface for events that can possibly have a related Person (#2021)
We have an `IEventWithPerson` interface that's used for extracting the
`PersonId` when we're saving events so we can populate the `person_id`
column. This column is indexed and exists for querying events by person
efficiently.

We have a few places coming up where events of the same type _might_
relate to a person but don't always. The `Guid PersonId` spec on
`IEventWithPerson` doesn't work in that case.

This adds `IEventWithOptionalPerson` where `PersonId` is nullable.

I considered making `IEventWithPerson` have a `Nullable<Guid>` instead
but it implies all our existing `IEventWithPerson` events can have null
`PersonId` set, which isn't valid.

Another alternative would be to use `Guid.Empty` or some other
well-known GUID but that seemed a little weird.