Recalculate professional status dates when a route is added or updated (#1996)
This adds a method to `Person` that recalculates the QTS/EYTS/PQTS/EYPS
attributes on Person when a given `ProfessionalStatusType` is changed.
This method is called when a route is added or updated.

I've updated the `ProfessionalStatus{Created|Updated}Event`s too to
capture previous and new QTS/EYTS/PQTS/EYPS attributes.

I've fixed up the `ProfessionalStatusTypes` for non-QTS routes in the
route reference data too.