Proactively sync matched Person before comparing attributes (#2023)
The process to resolve a TRN request uses Person data from the TRS DB.
This is synced in _near_ real time but can lag. Showing stale data on
these pages, particularly when the user is choosing what person data to
use, is not ideal.

This adds a proactive sync to GET requests for the pages that compare
the matched record to the request and show the resolved person
attributes. If the sync fails for whatever reason we won't block
progress; we don't want a transient error or CRM slow-ness to prevent
the journey being completed when 99% of time the data won't have
changed.

Once this is live we can monitor and adjust when/where we do this sync
and whether the 5 second threshold is sufficient.

We don't need a sync on the PersonDetail page (that we link to once the
request is resolved) since that's getting Person data from CRM so will
never be showing stale data.