Add pagination to list of API TRN requests (#2029)
I've tweaked the pagination bits added for the users list:
- The `Pagination` class is gone in favour of a `GetPageAsync()`
extension method on `IQueryable` and `ResultPage<T>`. With this
pagination can be chained onto an existing query.
- The `PaginationViewModel` now takes in a delegate that returns the
link so we can keep link generation in `TrsLinkGenerator`.

It would be good to have a reusable way to test pages that use these
pagination bits but I've left that for a future PR.