Wait for background job to finish before rendering list of tasks (#2036)
When an API TRN request is resolved, the update to the DQT record is
done in a background job. It's somewhat likely that a user will navigate
to the created/updated record immediately after resolving the task so
they may see stale data if the job hasn't yet completed.

This change passes the background job ID to the page that renders the
link to the created/updated record so that it can wait for a bit for the
job to complete. It'll only wait for up to 5 seconds; that should be
sufficient most of the time.