Add person matching for TRN requests on PersonMatchingService (#2040)
This adds `MatchFromTrnRequestAsync` and
`GetSuggestedMatchesFromTrnRequestAsync` to `PersonMatchingService` for
use in TRN requests. The rules are the same as today; a single match on
NINO and DOB is a definite match, matching on at least 3 of first name,
middle name, last name, email address, date of birth and NINO is a
potential match.

Note this is missing matching on previous names, since we don't have
those in TRS yet. A separate PR will cover that.