Allow job to run from a starting point for dqt annotations (#2048)
Backfilling notes takes a long time and sometimes times out. When the
job restarts, it starts from the beginning again. This PR allows for the
job to be restarted from the most recent row created in the notes table.

In the reporting database the oldest annotation was created on
19/03/2014, so i have defaulted it to that when there are no rows the
notes table.