Implement previous names (#2044)
### Context

We need to enable TRS to store and process Previous Names. This ticket
will cover the creation of the required tables to hold this information
and how this will interact with the audit trail.

### Changes proposed in this pull request

* Creates `PreviousName` entity and mapping
* Adds previous name whenever person details are edited and name is
changed
* Displays previous names on the person detail page
* **Note:** does not make a distinction between a name correction and a
formal name change as we don't have a UI for that yet
* This new behaviour is behind the `ContactsMigrated` feature flag, if
the flag is not enabled, the previous names are read/written to DQT

### Checklist

-   [x] Attach to Trello card
-   [x] Rebased master
-   [x] Cleaned commit history
-   [x] Tested by running locally
-   [ ] Run DQT integration tests locally (if appropriate)