Handle badly-formatted mobile numbers in sync (#2071)
Live sync is currently broken due to a phone number that's more than 15
characters long. From a quick scan of the reporting DB there's a lot of
bad data in the mobile number field.

This change will ensure we're not getting mobile numbers that are too
long. We'll have to re-visit what to do about the bad values when we get
to migrating the contact record.