Add a job to resync all Persons (#2080)
We've added a fair few fields to the Person sync recently and we need to
back-fill them. This adds a job that runs through all `Person`s already
in TRS DB and resyncs them. There's a date filter passed in so any that
have been synced since the last set of fields were added to the sync can
be excluded (and so that it can retry without re-processing rows).