Person qualification status summary (#2099)
### Context

Show any professional status that is held by the person on the person
details page, with a separate card

### Changes proposed in this pull request
https://trello.com/c/EgNZYbXA/1101-routes-add-summary-card-to-home-tab

### Guidance to review
card shows all 4 professional statuses and whether status is "Hold" or
"No" (not specified what it should say when it's not hold)
card shows extra row for each status that is held, with the held date. 
Content is not as described in ticket but instead, e.g.
Qualified teacher status            Holds
QTS holds date                           27 November 2023

---------

Co-authored-by: James Gunn <james@gunn.io>