Overhaul how we return a different API response based on role (#2102)
The `/v3/persons/<trn>` endpoint returns a different
`initialTeacherTraining` object if the calling user has the
`AppropriateBody` role. The way we did this before wan't great; we
wrapped every property in `Option<>` then had to apply custom type
converters to get things to AutoMap.

This changes that process to use a discriminated union type (`OneOf` in
this case).