Add GET /v3/trns/<trn> endpoint (#2117)
This adds a simple endpoint at `/v3/trns/<trn>` that can be used to
check if a TRN is valid. It returns a 204 if the record exists and is
active, a 308 (and a `Location` header) if the record has been merged
with another, 400 if the record is deactivated (but not merged) and 404
otherwise.


https://trello.com/c/vlc7XGe8/1259-api-updates-for-accessing-merged-records-via-the-api