Fix resync ignoring already-synced records (#2134)
Our resync job is having most records ignored because of the guard we
have in the sync process that prevents out-of-order updates. This
removes that check. (We don't really need the out-of-order check -
things will sort themselves out in the case that updates get replayed.)

I've increased the batch size too.