Disable API endpoints that write routes via configuration (#2138)
During routes migration we want to be sure we're not getting any route
writes coming through the API. Although we should have a block in place
in CRM this gives us an extra layer of protection.

I've tested this locally but not added automated tests as we can't
easily override configuration there currently.