Implement mapping logic from ITT / QTS -> TRS and output report (#2155)
Added mapping functions to `TrsDataSyncHelper` to map from ITT / QTS
records in DQT to `RouteToProfessionalStatus` in TRS.
This uses various sets of logic as defined in a spreadsheet in order to
derive the `RouteToProfessionalStatusType` to represent ITT / QTS as it
can be derived based on programme type , teacher status, qualification
and combinations of all of them.

This PR contains a CLI `route-mapper` which can be used to loop through
all active ITT / QTS for all contacts and output a CSV report detailing
how they will be mapped in TRS.

A further PR will apply the mappings to update the TRS database.

---------

Co-authored-by: James Gunn <james@gunn.io>