Add Gender field to TRS Person, including sync and backfill (#2179)
### Context

Following the decision to restructure the gender field for TRS we
decided that the Gender value held in DQT will be migrated to TRS and
held as a legacy value in case it is useful for future analysis.

This PR implements both of the following tickets:

https://trello.com/c/ZOm88D41/1314-create-gender-field-in-trs-sync-backfill

https://trello.com/c/iVU31lH8/1315-updates-to-the-create-record-journey-to-include-gender-field

### Changes proposed in this pull request

* Create `Gender` field on Person entity (and corresponding column on
`persons` table)
* Add `gender` column to `trs_persons` table in reporting DB
* Add `Gender` field to `TrsSyncHelper`
* Add logic and request validation to not allow `NotAvailable` to be
selected as a value for gender, but allow if it's a pre-existing value
that's left unchanged by the user.

I had to implement the sync as the tests are dependent on the sync
process to function.

### Checklist

-   [x] Attach to Trello card
-   [x] Rebased master
-   [x] Cleaned commit history
-   [x] Tested by running locally
-   [ ] Run DQT integration tests locally (if appropriate)