Added event when route is migrated (#2191)
# Context
ITT/QTS are moving into TRS as ‘route to professional status’.
A background job has been implemented which writes data to the
qualifications table in TRS.
We also need to output an event in TRS for each route so the audit
history shows how they were mapped from DQT to TRS.

# Details
Create a new RouteToProfessionalStatusMigratedEvent which contains all
of the values from ITT / QTS used to derive the route info in TRS along
with the TRS route values.
Amend the code so that the events are written to the db as part of the
migration.