Added RouteMigrationReportItem model and mapping (#2197)
### Context

ITT/QTS are moving into TRS as ‘route to professional status’.

A background job has been implemented which writes data to the
qualifications table in TRS.

We also need to write out a report of how the mappings were derived to a
new table to allow us to verify that they were mapped from DQT to TRS
correctly.

### Changes proposed in this pull request

Create a new RouteMigrationMappingReport model to store the same
information as the original mapping CSV report. Ensure that this is
added to the reporting database too. This can then be used as an
additional source to verify how the mappings were derived (containing
more detail than the event).
Amend the code so that the report is written to the db as part of the
migration.

Additionally a bug fix where QTLS exemption reasons were not being moved
from Person to Route has now been fixed