Rewire TRN request API to create support tasks in TRS (#2051)
This updates the TRN request endpoints to create support tasks in TRS
when a duplicate review is required. I've removed/simplified some of the
API integration tests where the behaviour being tested is now covered by
unit tests.


`TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/Implementation/Operations/CreateTrnRequest.cs`
is where the bulk of the interesting changes are.

The trickiest part of this was wrangling the background job execution in
tests; we have a new `ExecuteOnCommitBackgroundJobScheduler` that will
execute any scheduled jobs at the point the current `Transaction` is
committed. Using `ExecuteImmediatelyBackgroundJobScheduler` won't work
here since the data the job needs won't be available to the job
'immediately' (since the transaction isn't committed yet).

As part of this I've also moved the
`ExecuteImmediatelyBackgroundJobScheduler` into `TestCommon`.