Rationalize events that create or update a Person (#2290)
We have support tasks that, when resolved, create or update a person. We
need a single event to represent each of those actions that covers both
the person update (or creation) and resolving the task. Our old pattern
of `Person.Create()`, `Person.UpdateDetails()` etc. that updates the
relevant model properties and creates an event doesn't work here; we
need different event types depending on the context.

This modifies `Person.Create()` and `Person.UpdateDetails()` to *not*
create events, but instead to return a `Result` type that allows the
caller to create the event. In addition, `IEventWithPersonAttributes` is
added so we can easily find event types that have created or updated a
`Person` and query them consistently.

At some point we should review all our model methods so we're consistent
with where the events are created and perhaps move some of this logic
into Service classes.