Add support for a secondary person ID in events (#2307)
We have some events, notably `PersonsMergedEvent`, that relate to
multiple Persons so our scalar `PersonId` event doesn't work. This adds
a `PersonIds` property, an `IEventWithSecondaryPersonId` interface and
updates the event mapping to populate it.

Subsequent PRs will follow to back-fill `PersonId`s, start using it, and
finally to remove `PersonId`.