Amended DOB change to create support task in TRS (behind feature flag) (#2302)
### Context

The TRS Console currently allows users to manage Change name and change
date of birth requests which are stored as “Cases” in DQT.
As part of migrating from DQT, we need to move these to be stored as
support tasks in TRS.

### Changes proposed in this pull request

Create a new record `ChangeDateOfBirthRequestData` in the
`TeachingRecordSystem.Core.Models.SupportTaskData` namespace based on
existing fields in `CreateDateOfBirthChangeRequestCommand` replacing
`EvidenceFileUrl` with `EvidenceFileId`.

Amend `CreateDateOfBirthChangeRequestHandler` to create a `SupportTask`
of type `ChangeDateOfBirthRequest` populating the `Data` field with a
`ChangeDateOfBirthRequestData` record populated from the incoming
request (when the feature flag is enabled otherwise call the existing
code which creates cases in DQT).

Considerations:
- The incoming request populates the `EvidenceFileUrl` with an expiring
link to download uploaded evidence.
We need to use this URL to download the evidence and then re-upload it
to the `uploads` container in Azure Blob Storage for TRS, storing the
file ID returned from blob storage in the
`ChangeDateOfBirthRequestData.EvidenceFileId` field.
- We need to replicate the call to send an email from the CRM plugin
`CaseCreatedSendEmailPlugin` (using `INotificationSender` to send emails
via Notify in TRS).
- This should be kept behind a feature flag.