Added support for creating name change support tasks in TRS (#2318)
### Context

The TRS Console currently allows users to manage Change name and change
date of birth requests which are stored as “Cases” in DQT.
As part of migrating from DQT, we need to move these to be stored as
support tasks in TRS.

### Changes proposed in this pull request

Create a new record `ChangeNameRequestData` in the
`TeachingRecordSystem.Core.Models.SupportTaskData` namespace based on
existing fields in `CreateNameChangeRequestCommand` replacing
`EvidenceFileUrl` with `EvidenceFileId`.
Amend `CreateNameChangeRequestHandler` (use a feature flag to keep the
existing DQT related functionality toggleable) to create a `SupportTask`
of type `ChangeNameRequest` populating the Data field with a
`ChangeNameRequestData` record populated from the incoming request.

Considerations:
- The incoming request populates the `EvidenceFileUrl` with an expiring
link to download uploaded evidence.
We need to use this URL to download the evidence and then re-upload it
to the uploads container in Azure Blob Storage for TRS, storing the file
ID returned from blob storage in the
`ChangeNameRequestData.EvidenceFileId` field..
- We need to replicate the call to send an email from the CRM plugin
`CaseCreatedSendEmailPlugin` (using `INotificationSender` to send emails
via Notify in TRS).
- This should be kept behind a feature flag UI to manage the support
tasks is in production.